refactor: 입금 관리 알림 UI를 PaymentAlertCard로 교체
